Переписывает приложение
