--- src/const.js
@@ -1,3 +1,0 @@
-const POINTS_TYPE = ['Taxi', 'Ship', 'Flight'];
-
-export { POINTS_TYPE };
