+++ src/view/container-points-view.js
@@ -1,0 +1,13 @@
+import AbstractView from '../framework/view/abstract-view';
+
+function createContainerForPoints() {
+  return /*html*/ `
+    <ul class="trip-events__list"></ul>
+  `;
+}
+
+export default class ContainerForPoints extends AbstractView {
+  get template() {
+    return createContainerForPoints();
+  }
+}
