+++ src/view/point-view.js
@@ -1,47 +1,24 @@
-import AbstractView from '../framework/view/abstract-view.js';
+import AbstractView from '../framework/view/abstract-view';
 import {
-  filterHoursPoints,
   getTimeInterval,
+  filterHoursPoints,
   filterDayMonth,
-  filterPointDay,
-} from '../utils.js';
-import { mockOffers } from '../mock/points.js';
+  filterPointDay
+} from '../utils';
 
-function createBoardTemplate(point) {
-  const { typePoint, destination, timeStart, timeEnd, offersCheck, basePrice, isFavorite } = point;
+function createPoint(point, listOffers) {
+  const { basePrice, destination, isFavorite, offersCheck, timeStart, timeEnd, typePoint } = point;
 
-  const typeOffersObj = mockOffers.find((item) => item.type === typePoint);
+  const day = filterDayMonth(timeStart);
+  const hoursStart = filterHoursPoints(timeStart);
+  const hoursEnd = filterHoursPoints(timeStart);
+  const timeInterval = getTimeInterval(timeStart, timeEnd);
+  const startPointDay = filterPointDay(timeStart);
+
+  const typeOffersObj = listOffers.find((item) => item.type === typePoint);
   const { offers } = typeOffersObj;
 
-  const normalizeTimeStart = filterHoursPoints(timeStart);
-  const normalizeTimeEnd = filterHoursPoints(timeEnd);
-  const normalizeDate = filterDayMonth(timeStart);
-  const favoriteClassName = isFavorite ? 'event__favorite-btn--active' : '';
-
-  return /*html*/ `
-  <li class="trip-events__item">
-    <div class="event">
-      <time class="event__date" datetime="${filterPointDay(
-    timeStart
-  )}">${normalizeDate}</time>
-      <div class="event__type">
-      <img class="event__type-icon" width="42" height="42" src="img/icons/${typePoint}.png" alt="Event type icon">
-    </div>
-    <h3 class="event__title">${typePoint} ${destination}</h3>
-      <div class="event__schedule">
-        <p class="event__time">
-          <time class="event__start-time" datetime="2019-03-19T18:00">${normalizeTimeStart}</time>
-          &mdash;
-          <time class="event__end-time" datetime="2019-03-19T19:00">${normalizeTimeEnd}</time>
-        </p>
-      <p class="event__duration">${getTimeInterval(timeStart, timeEnd)}</p>
-    </div>
-    <p class="event__price">
-      &euro;&nbsp;<span class="event__price-value">${basePrice}</span>
-    </p>
-    <h4 class="visually-hidden">Offers:</h4>
-    <ul class="event__selected-offers">
-      ${offers.map(({ id, title, price }) => offersCheck.map((checkedOfferId) => {
+  const createTypeOffersTemplate = () => offers.map(({ id, title, price }) => offersCheck.map((checkedOfferId) => {
     if (checkedOfferId === id) {
       return `<li class="event__offer">
               <span class="event__offer-title">${title}</span>
@@ -49,37 +26,62 @@ function createBoardTemplate(point) {
               <span class="event__offer-price">${price}</span>
             </li>`;
     }
-  }).join('')).join('')}
-    </ul >
-    <button class="event__favorite-btn ${favoriteClassName}" type="button">
-      <span class="visually-hidden">Add to favorite</span>
-      <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
-        <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
-      </svg>
-    </button>
-    <button class="event__rollup-btn" type="button">
-      <span class="visually-hidden">Open event</span>
-    </button>
-  </div >
-</li > `;
+  }).join('')).join('');
+
+
+  const favoriteClassPoint = isFavorite ? 'event__favorite-btn--active' : '';
+
+  return /*html*/ `
+    <li class="trip-events__item">
+      <div class="event">
+        <time class="event__date" datetime="${startPointDay}">${day}</time>
+        <div class="event__type">
+          <img class="event__type-icon" width="42" height="42" src="img/icons/${typePoint}.png" alt="Event type icon">
+        </div>
+        <h3 class="event__title">${typePoint} ${destination}</h3>
+        <div class="event__schedule">
+          <p class="event__time">
+            <time class="event__start-time" datetime="${timeStart}">${hoursStart}</time>
+            &mdash;
+            <time class="event__end-time" datetime="${timeEnd}">${hoursEnd}</time>
+          </p>
+          <p class="event__duration">${timeInterval}</p>
+        </div>
+        <p class="event__price">
+          &euro;&nbsp;<span class="event__price-value">${basePrice}</span>
+        </p>
+        <h4 class="visually-hidden">Offers:</h4>
+        <ul class="event__selected-offers">
+
+
+        ${ createTypeOffersTemplate(offers) }
+
+        </ul>
+        <button class="event__favorite-btn ${favoriteClassPoint}" type="button">
+          <span class="visually-hidden">Add to favorite</span>
+          <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
+            <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
+          </svg>
+        </button>
+        <button class="event__rollup-btn" type="button">
+          <span class="visually-hidden">Open event</span>
+        </button>
+      </div>
+    </li>
+  `;
 }
 
 export default class Point extends AbstractView {
-  #handleClick = null;
+  #point;
+  #listOffers;
 
-  constructor(point) {
+  constructor({ point, listOffers }) {
     super();
-    this.point = point;
-    // this.handleClick = onClick;
-    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#clickHandler);
+    this.#point = point;
+    this.#listOffers = listOffers;
   }
 
   get template() {
-    return createBoardTemplate(this.point);
+    return createPoint(this.#point, this.#listOffers);
   }
-
-  #clickHandler = (evt) => {
-    evt.preventDefault();
-    // this.handleClick();
-  };
 }
