--- src/view/route-view.js
@@ -1,14 +1,0 @@
-import AbstractView from '../framework/view/abstract-view.js';
-
-function createBoardTemplate() {
-
-  return /*html*/ `
-  <ul class="trip-events__list">
-  </ul>`;
-}
-
-export default class RouteTrip extends AbstractView {
-  get template() {
-    return createBoardTemplate();
-  }
-}
